+++ src/app.ts
@@ -2,10 +2,12 @@ import express from 'express';
 import bodyParser from 'body-parser';
 import proxyController from "./controller/ProxyController";
 import { LoggerMiddleware } from './middleware/LoggingMiddleware';
-import { myDataSource } from './config/database.config';
+import { dataSource } from './config/database.config';
 import userController from './controller/user/UserController';
+import superAdminController from './controller/user/SuperAdminController';
+import tenantController from './controller/user/TenantController';
 
-myDataSource
+dataSource
   .initialize()
   .then(() => {
     console.log("Data Source has been initialized!")
@@ -21,5 +23,7 @@ app.use(LoggerMiddleware.logRequest)
 app.use(LoggerMiddleware.logResponseTime)
 app.use('/proxy', proxyController);
 app.use('/user', userController);
+app.use('/superadmin', superAdminController);
+app.use('/tenant', tenantController);
 
 export default app;
