+++ src/repository/document.repository.ts
@@ -1,9 +1,9 @@
-import { myDataSource } from '../config/database.config';
+import { dataSource } from '../config/database.config';
 import { DocumentEntity } from '../entity/document.entity';
 
 export class DocumentRepository {
   private readonly documentRepository =
-    myDataSource.getRepository(DocumentEntity);
+    dataSource.getRepository(DocumentEntity);
 
   public async save(document: DocumentEntity): Promise<DocumentEntity> {
     return await this.documentRepository.save(document);
